% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{a_star}
\alias{a_star}
\title{Calculate a*}
\usage{
a_star(dat)
}
\arguments{
\item{dat}{\verb{[data.frame]} The location data used for fitting the LoCoH.
Should contain the following columns:
\itemize{
\item \verb{$x} -- The x-coordinate of the animal's location
\item \verb{$y} -- The y-coordinate of the animal's location
\item \verb{$t} -- The date and time (as a \code{POSIXct} object) of the location
}}
}
\value{
A numeric vector of length 1 with the value of a*
}
\description{
Calculates suggested starting value for \code{a} under a-LoCoH
}
\details{
Calculates a* as the maximum distance between any two points in the
dataset.
}
\examples{

data(tracks)

winter01 <- tracks[which(tracks$id == "ID01" & tracks$season == "Winter"), ]

a_star(winter01)

}
\seealso{
Getz et al. (2007) \emph{insert full citation here}
}
