% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{bcrw}
\alias{bcrw}
\title{Simulate Biased Correlated Random Walk}
\usage{
bcrw(
  start_loc,
  centroid,
  n_steps,
  sl_distr = c(shape = 0.75, scale = 200),
  rho = 0.25,
  beta = 0.1
)
}
\arguments{
\item{start_loc}{\verb{[numeric]} Vector of length 2 giving x and y coordinates
of starting location.}

\item{centroid}{\verb{[numeric]} Vector of length 2 giving x and y coordinates of
home range centroid.}

\item{n_steps}{\verb{[integer]} Integer giving number of steps to simulate.}

\item{sl_distr}{\verb{[numeric]} \emph{Named} vector of length 2 giving shape and
scale of gamma distribution from which step lengths are drawn.}

\item{rho}{\verb{[numeric]} Value between 0 and 1 giving correlation of turning
angles.}

\item{beta}{\verb{[numeric]} Value between 0 and 1 giving the strength of the
attraction towards the centroid, \emph{i.e.}, strength of the bias.}
}
\description{
Simulate Biased Correlated Random Walk
}
\author{
Brian J. Smith & Simona Picardi
}
