% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{calc_emd}
\alias{calc_emd}
\title{Calculate EMD}
\usage{
calc_emd(r_list, mat, check_symmetric = TRUE)
}
\arguments{
\item{r_list}{\verb{[list]} List of \code{r} rasterized home ranges to compare.}

\item{mat}{\verb{[matrix]} \code{r} x \code{r} matrix of indicating whether a row-column
pair should have its EMD calculated. Can be logical or integer(ish). EMD
will be calculated for pairs where \code{mat == TRUE} or \code{mat == 1}. See details.}

\item{check_symmetric}{\verb{[logical = TRUE]} Controls whether function will check for
\code{mat} for symmetry. See details.}
}
\description{
Calculates Earth-mover's distance for a list of rasters
}
\details{
EMD calculations are symmetric, i.e., \code{emd(x, y) == emd(y, x)}.
These calculations are time-consuming, so you will probably want to avoid
providing a symmetric matrix. The argument \code{check_symmetric} defaults to
\code{TRUE} and causes the function to stop if your matrix is symmetric so you
can correct this before proceeding.
}
\examples{
\dontrun{

# Load data
data(tracks)

# Keep only winter data
winter <- tracks[which(tracks$season == "Winter"), ]

# Split by ID
winter_list <- split(winter, winter$id)

# Calculate a for a-LoCoH
a <- a_star(winter_list)

# Fit LoCoHs
locoh_list <- lapply(winter_list, fit_locoh, n = a)

# Rasterize
r_list <- lapply(locoh_list, rasterize_locoh, res = 500)

# Matrix of comparisons
comp <- matrix(0,
nrow = length(winter_list),
ncol = length(winter_list))

comp[1, 2] <- comp[2, 3] <- comp[3, 4] <- 1

# Calculate EMD
EMD <- calc_emd(r_list, mat = comp)
}

}
