% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_moving_window.R
\name{comp_mat.mv_wind}
\alias{comp_mat.mv_wind}
\title{Comparison Matrix for Moving Window}
\usage{
comp_mat.mv_wind(x)
}
\arguments{
\item{x}{An \code{mv_wind} object returned by
\code{\link{construct_windows}()}.}
}
\description{
Makes a comparison matrix for the moving window analysis
}
\examples{

# Load lubridate
library(lubridate)

# Grab a date/time
right_now <- Sys.time()

# Construct windows
wind <- construct_windows(start = right_now,
                          width = days(10),
                          step = days(10),
                          end = right_now + days(100))

# Create comparison matrix
mat <- comp_mat.mv_wind(wind)

}
