% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_moving_window.R
\name{construct_windows}
\alias{construct_windows}
\title{Construct time windows}
\usage{
construct_windows(start, width, step, end)
}
\arguments{
\item{start}{\verb{[POSIXt]} The start date and time for the first window.}

\item{width}{\verb{[Period]} A \code{\link[lubridate:period]{Period}} object
specifying the width of each moving window.}

\item{step}{\verb{[Period]} A \code{\link[lubridate:period]{Period}} object
specifying the space between start times for each moving window.}

\item{end}{\verb{[POSIXt]} The last date and time to include at the end of a
window.}
}
\value{
Returns a \code{data.frame} with columns:
\itemize{
\item \code{window} An integer indexing the window number
\item \code{start} The date and time of the window start (class \code{POSIXct})
\item \code{end} The date and time of the window end (class \code{POSIXct})
}
}
\description{
Constructs time windows for moving window analysis
}
\examples{

# Load lubridate
library(lubridate)

# Grab a date/time
right_now <- Sys.time()

# Construct windows
construct_windows(start = right_now,
                  width = days(5),
                  step = days(1),
                  end = right_now + days(100))

}
