% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locoh.R
\name{fit_locoh}
\alias{fit_locoh}
\title{Fit one LoCoH home range}
\usage{
fit_locoh(dat, crs, type, n, ...)
}
\arguments{
\item{dat}{\verb{[data.frame]} The location data used for fitting the LoCoH.
Should contain the following columns:
\itemize{
\item \verb{$x} -- The x-coordinate of the animal's location
\item \verb{$y} -- The y-coordinate of the animal's location
\item \verb{$t} -- The date and time (as a \code{POSIXct} object) of the location
}}

\item{crs}{\verb{[sp::CRS = NULL]} The (optional) coordinate reference system for
the location data.}

\item{type}{\verb{[character = "a"]} The type of neighbor rule used to create the
local hulls. Should be one of \code{"a"} (default), \code{"k"}, or \code{"r"}.}

\item{n}{\verb{[numeric = NA]} The value to use for the neighbor rule, \emph{i.e.}, the
value of either a, k, or r used to fit the local hulls.}

\item{`...`}{Additional arguments passed to \code{\link{amt::hr_locoh}()}.}
}
\description{
Fits a single LoCoH home range to data
}
\details{
A wrapper for \code{amt} functions to fit LoCoH home ranges. First
creates a \code{track_xyt} object, then calls \code{\link{amt::hr_locoh}()}.
}
\examples{
\dontrun{
# Example
}
}
\seealso{
\code{\link{amt::hr_locoh}()}
Getz et al. 2007 (\emph{insert full citation here})
}
