% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoCoHverlap.R
\docType{package}
\name{LoCoHverlap}
\alias{LoCoHverlap}
\title{LoCoHverlap: A package for measuring animal range overlap}
\description{
More documentation to follow
}
\section{Basic workflow}{

The basic functions ...
}

\section{Seasonal workflow}{

The seasonal functions ...
}

\section{Moving window workflow}{

The moving window functions ...
}

