% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_moving_window.R
\name{mw_analysis}
\alias{mw_analysis}
\title{LoCoHverlap Moving Window Analysis}
\usage{
mw_analysis(
  dat,
  start,
  width,
  step,
  end,
  crs = sf::NA_crs_,
  type = "a",
  n = a_default(dat),
  res = 500,
  verbose = TRUE
)
}
\arguments{
\item{dat}{\verb{[data.frame | list]} The location data used for fitting the
LoCoH. Can be a single \code{data.frame} or a list of \code{data.frame}s with the
following columns:
\itemize{
\item \verb{$x} -- The x-coordinate of the animal's location
\item \verb{$y} -- The y-coordinate of the animal's location
\item \verb{$t} -- The date and time (as a \code{POSIXct} object) of the location
}}

\item{start}{\verb{[POSIXt]} The start date and time for the first window.}

\item{width}{\verb{[Period]} A \code{\link[lubridate:period]{Period}} object
specifying the width of each moving window.}

\item{step}{\verb{[Period]} A \code{\link[lubridate:period]{Period}} object
specifying the space between start times for each moving window.}

\item{end}{\verb{[POSIXt]} The last date and time to include at the end of a
window.}

\item{crs}{\verb{[sf::crs = sf::NA_crs_]} The (optional) coordinate reference
system for the location data. May also be passed as an integer representing
the EPSG code.}

\item{type}{\verb{[character = "a"]} The type of neighbor rule used to create the
local hulls. Should be one of \code{"a"} (default), \code{"k"}, or \code{"r"}.}

\item{n}{\verb{[numeric = NA]} The value to use for the neighbor rule, \emph{i.e.}, the
value of either a, k, or r used to fit the local hulls.}

\item{res}{\verb{[numeric = 500]} The resolution of the \code{RasterLayer} to create
as the template raster. Units should be in the units of the CRS, \emph{e.g.} if
\code{proj=utm}, units should be in meters and if \code{proj=longlat}, units should be
in degrees. Defaults to 500 (assumes meters).}

\item{verbose}{\verb{[logical = TRUE]} Should the function report what step it
is on? Defaults to \code{TRUE} and cats updates as it works.}
}
\description{
Runs a full moving window analysis on a dataset.
}
\details{
Current implementation does not accept a custom raster like \code{fit_locoh()}
does.
}
