% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_moving_window.R
\name{nest_windows}
\alias{nest_windows}
\title{Nest data into windows}
\usage{
nest_windows(dat, mv_wind)
}
\arguments{
\item{dat}{\verb{[data.frame]} A \code{data.frame} with location data for one
individual. Should have same columns as \code{dat} for \code{\link{fit_locoh}()}.}

\item{mv_wind}{\verb{[mv_wind]} An \code{mv_wind} object returned by
\code{\link{construct_windows}()}.}
}
\description{
Creates a nested \code{data.frame} from an \code{mv_wind} object
}
\details{
Uses \code{\link[tidyr:nest]{nest}()} and
\code{\link[sqldf:sqldf]{sqldf}()}
}
\examples{
\dontrun{

# Load lubridate
library(lubridate)

# Load data
data(tracks)

# Keep 1 individual
tk1 <- tracks[which(tracks$id == "ID01"), ]

# Construct moving windows
wind <- construct_windows(start = ymd_hms("2021-05-15 00:00:00"),
                          width = days(5),
                          step = days(2),
                          end = ymd_hms("2021-06-30 00:00:00"))

#Nest
nst <- nest_windows(tk1, wind)

}

}
