% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_moving_window.R
\name{plot.mv_wind}
\alias{plot.mv_wind}
\title{Plot moving windows}
\usage{
\method{plot}{mv_wind}(x, set_par = TRUE, ...)
}
\arguments{
\item{x}{\verb{[mv_wind]} An \code{mv_wind} object returned by
\code{\link{construct_windows}()}.}

\item{set_par}{\verb{[logical = TRUE]} Should the function be allowed to set
its own graphical parameters?}

\item{...}{Arguments passed to \code{segments()}.}
}
\description{
Visualizes moving windows from \code{mv_wind} object
}
