% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_moving_window.R
\name{plot.mwa}
\alias{plot.mwa}
\title{Plot an \code{mwa} Object}
\usage{
\method{plot}{mwa}(x, set_par = TRUE, log = TRUE, ...)
}
\arguments{
\item{x}{\verb{[mwa]} An object of class \code{mwa} returned by
\code{\link{mw_analysis}()}.}

\item{set_par}{\verb{[logical = TRUE]} Should the function be allowed to set
its own graphical parameters?}

\item{log}{\verb{[logical = TRUE]} Should the function plot \code{log(emd)} (default)
or just \code{emd}?}

\item{...}{Arguments passed to \code{plot.default()}.}
}
\description{
Plots an object of class \code{mwa}.
}
