% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{rasterize_locoh}
\alias{rasterize_locoh}
\title{Create a UD from a LoCoH}
\usage{
rasterize_locoh(locoh, r, res)
}
\arguments{
\item{locoh}{\verb{[locoh]} The fitted LoCoH returned by
\code{\link{fit_locoh}()}.}

\item{r}{\verb{[RasterLayer]} The template raster to estimate the UD on.
If not specified, a raster will be created using the bounding box of \code{locoh}
and the resolution specified by \code{res}.}

\item{res}{\verb{[numeric]} The resolution of the \code{RasterLayer} to create as the
template raster. Ignored if \code{rast} is provided.}
}
\description{
Converts a fitted LoCoH to a utilization distribution
}
