% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_moving_window.R
\name{nest_windows}
\alias{nest_windows}
\title{Nest data into windows}
\usage{
nest_windows(dat, mv_wind)
}
\arguments{
\item{dat}{\verb{[data.frame]} A \code{data.frame} with location data for one
individual. Should have same columns as \code{dat} for \code{\link{fit_locoh}()}.}

\item{mv_wind}{\verb{[mv_wind]} An \code{mv_wind} object returned by
\code{\link{construct_windows}()}.}
}
\value{
Returns a nested \code{data.frame} (a \code{tbl}) with columns:
\itemize{
\item \code{window} -- An integer indexing the window number
\item \code{start} -- The date and time of the window start (class \code{POSIXct})
\item \code{end} -- The date and time of the window end (class \code{POSIXct})
\item \code{data} -- A list column containing \code{tbl}s with the animal location
data. It has columns:
\itemize{
\item \code{t} -- The location timestamp (class \code{POSIXct})
\item \code{x} -- The x-coordinate of the location
\item \code{y} -- The y-coordinate of the location
}
}
}
\description{
Creates a nested \code{data.frame} from an \code{mv_wind} object
}
\details{
Uses \code{\link[sqldf:sqldf]{sqldf}()} to join locations to windows
that they fall within. Then uses \code{\link[tidyr:nest]{nest}()} to create
a nested \code{data.frame} holding the location data in a list column.
}
\examples{
\dontrun{

# Load lubridate
library(lubridate)

# Load data
data(tracks)

# Keep 1 individual
tk1 <- tracks[which(tracks$id == "ID01"), ]

# Construct moving windows
wind <- construct_windows(start = ymd_hms("2021-05-15 00:00:00"),
                          width = days(5),
                          step = days(2),
                          end = ymd_hms("2021-06-30 00:00:00"))

#Nest
nst <- nest_windows(tk1, wind)

}

}
